requirements.txt
asgiref==3.3.4
coverage==5.5
deprecation==2.1.0
Django==3.2.4
eyed3==0.9.6
filetype==1.0.7
packaging==20.9
pyparsing==2.4.7
python-Levenshtein==0.12.2
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.10.0.0
youtube-dl==2021.6.6
